Fix updating of cache
